dynamic for code generation

+++ MIPT_Compiler/Analyzer/VirtualTable.h
@@ -16,7 +16,7 @@ public:
 	int GetMethodIndex(const CSymbol* name) const;
 	void ReplaceMethod(const CFunctionInfo* method);
 	void AddMethod(const CFunctionInfo* method) { methods.push_back(method); };
-	int GetSize() const { return methods.size(); };
+	int GetSize() const { return static_cast<int>(methods.size()); };
 	std::shared_ptr<const IR::CNameExpression> GetTableName() const;
 
 private:
